refactor: extract aux streaming cluster into agent/aux_streaming.py (R5 C-B)
Move the progress-hooked chat.completions streaming machinery out of the
agent/auxiliary_client.py godfile (window 8221-8558) into a dedicated
agent/aux_streaming.py module: _AUX_STREAM_CEILING_* constants,
_aux_stream_total_ceiling, _client_streams_internally,
_is_streaming_rejected_error, _provider_requires_stream,
_create_with_progress, _aggregate_chat_stream, _ChatStreamAccumulator,
_aggregate_chat_stream_async and _acreate_with_stream.

The godfile keeps an identity re-export block so every moved name resolves
is-identical through agent.auxiliary_client (callers, the compression
pipeline, and tests keep importing from the godfile unchanged). Godfile-owned
dependencies (logger, wire-adapter client classes, error predicates,
_aux_progress_active/_notify_aux_progress) are imported lazily inside the
moved functions to keep the module import-order independent.

Adds tests/test_aux_streaming_seam.py asserting real identity for all 11
re-exported names plus accumulator/stream-aggregation smoke coverage.

Epic #78647, target #78635. Zero behavior change.
